+++ index.js
@@ -1,15 +1,18 @@
 'use strict';
 
 /**
- * Checks if a specifier is an extensionless Babel runtime specifier.
- * @param {string} specifier Specifier to check.
- * @returns {boolean} Is the specifier an extensionless Babel runtime specifier.
+ * Ensures a require or import string literal specifier has a file extension if
+ * it’s for the Babel runtime.
+ * @param {{value: string}} specifier Require or import string literal specifier.
  */
-function isExtensionlessBabelRuntimeSpecifier(specifier) {
-  return (
-    specifier.startsWith('@babel/runtime/helpers/') &&
-    !specifier.endsWith('.js')
-  );
+function fixBabelRuntimeSpecifier(specifier) {
+  if (
+    specifier.value.startsWith('@babel/runtime/helpers/') &&
+    !specifier.value.endsWith('.js')
+  )
+    specifier.value = `${specifier.value}.js`;
+  else if (specifier.value === '@babel/runtime/regenerator')
+    specifier.value = `${specifier.value}/index.js`;
 }
 
 /**
@@ -27,11 +30,7 @@ module.exports = function babelPluginTransformRuntimeFileExtensions() {
       // Example ExportNamedDeclaration (unlikely to occur):
       //   export { default as _objectWithoutPropertiesLoose } from "@babel/runtime/helpers/objectWithoutPropertiesLoose";
       'ImportDeclaration|ExportNamedDeclaration'(path) {
-        if (
-          path.node.source &&
-          isExtensionlessBabelRuntimeSpecifier(path.node.source.value)
-        )
-          path.node.source.value = `${path.node.source.value}.js`;
+        if (path.node.source) fixBabelRuntimeSpecifier(path.node.source);
       },
 
       // Example CallExpression:
@@ -39,12 +38,8 @@ module.exports = function babelPluginTransformRuntimeFileExtensions() {
       CallExpression(path) {
         if (path.node.callee.name === 'require') {
           const [specifier] = path.node.arguments;
-          if (
-            specifier &&
-            specifier.type === 'StringLiteral' &&
-            isExtensionlessBabelRuntimeSpecifier(specifier.value)
-          )
-            specifier.value = `${specifier.value}.js`;
+          if (specifier && specifier.type === 'StringLiteral')
+            fixBabelRuntimeSpecifier(specifier);
         }
       },
     },